new api endpoints for GA data import
