+++ lib/analytics.js
@@ -548,3 +548,13 @@ export async function hasuraInsertReadingDepth(params) {
     },
   });
 }
+
+export function sanitizePath(path) {
+  if (/\?/.test(path)) {
+    path = path.split('?')[0];
+  }
+  if (/\/en-US\//.test(path)) {
+    path = path.replace('/en-US', '');
+  }
+  return path;
+}
